coach alert

+++ resources/views/tryouts/show.blade.php
@@ -35,13 +35,14 @@
 				<h1>No Additional Information Has Been Provided</h1>
 			@endif
 	@endforeach
+		<?php $date = date('Y-m-d'); ?>
 		@if(Auth::user())
-			@if(Auth::user()->id == $tryout->user_id)
+			@if((Auth::user()->id == $tryout->user_id) && ($tryout->date > $date))
 				<div>
 					<a href="/tryouts/{{$id}}/edit" class="btn btn-lg">Edit Post</a>
 					<a href="/tryouts/{{$id}}/delete" class="btn btn-lg" onclick="return confirm('Are you sure you want to delete this post?');">Delete Post</a>
 				</div>
-			@else
+			@elseif($tryout->date > $date)
 				{!! Form::model($tryout, ['method' => 'POST', 'class' => 'form-horizontal', 'action' => ['TryoutsController@rsvp', $tryout->id]]) !!}
 
 				<div class="form-group">
@@ -50,7 +51,7 @@
 				</div>
 			{!! Form::close() !!}
 			@endif
-		@else
+		@elseif($tryout->date > $date)
 			{!! Form::model($tryout, ['method' => 'POST', 'class' => 'form-horizontal', 'action' => ['TryoutsController@rsvp', $tryout->id]]) !!}
 
 				<div class="form-group">
